FolderTest 생성 및 FolderService 인터페이스 삭제와 FolderService 클래스 생성
