+++ src/test/java/com/mashup/molinkbackend/folder/FolderTest.java
@@ -1,0 +1,4 @@
+package com.mashup.molinkbackend.folder;
+
+public class FolderTest {
+}
